Refactor staffusers module and enhance user management
This commit majorly refactors the staffusers module to enhance functionality and user management. Changes include the creation of new 'borrow_book' and 'return_book' functions in staffuser views, the removal of the 'register' and 'create_book' paths, and addition of a staff_home view. Authentication and user creation have been improved as UserCreationForm and CreateBookForm have been introduced in admin.py. Lastly, the user interface of the login.html template has been updated for a better user experience.