Implemented book lending and reservation features
Created two new models in users app, BorrowedBook and Reservation, which track the current state of a borrowed or reserved book. Updated book_detail view and template to allow users to reserve, borrow, or cancel a reservation for a book. Adjusted staffuser's index, add, edit, and delete views and templates for handling new book management system. Optimized pagination and search views for staffuser and user pages.